Fix failing imports & pass tests
