+++ src/routes/MarketList/store/test/market.reducer.js
@@ -2,7 +2,7 @@ import { combineReducers, createStore, applyMiddleware, compose } from 'redux'
 import thunk from 'redux-thunk'
 import { List } from 'immutable'
 import marketReducer, { REDUCER_ID } from 'store/reducers/market'
-import blockchainReducer from 'reducers/blockchain'
+import blockchainReducer from 'store/reducers/blockchain'
 import { processMarketResponse } from '../actions/fetchMarkets'
 import { marketListSelector } from '../selectors'
 import { oneMarketData, twoMarketData, realData, MarketFactory } from './builder/index.builder'
