+++ src/store/middlewares/LocalStorageDump.js
@@ -1,21 +1,34 @@
+import { pick, set } from 'lodash'
+
 import { LOAD_LOCALSTORAGE } from './LocalStorageLoad'
 import { LOAD_SESSIONSTORAGE } from './SessionStorageLoad'
 
 const CLEAR_LOCAL_STORAGE = 'CLEAR_LOCAL_STORAGE'
+
 const INIT = 'INIT'
 const forbiddenActions = [INIT, CLEAR_LOCAL_STORAGE, LOAD_LOCALSTORAGE, LOAD_SESSIONSTORAGE]
 
-const PERSIST_PATHS = ['transactions']
+const PERSIST_PATHS = [
+  'transactions.log',
+]
 
 export default store => next => (action) => {
   const state = store.getState()
+
   if (forbiddenActions.indexOf(action.type) === -1) {
-    const storage = {}
+    let storage = {}
 
     PERSIST_PATHS.forEach((path) => {
-      storage[path] = { ...state[path].toJS() }
+      storage = {
+        ...pick(state, path),
+      }
     })
 
+    // FIX-ME: Temporary to remember ToS Acceptance
+    const tosAccepted = state.integrations.get('termsAndConditionsAccepted').toJS()
+    set(storage, 'integrations.termsAndConditionsAccepted', tosAccepted)
+
+    // eslint-disable-next-line no-undef
     window.localStorage.setItem(`GNOSIS_${process.env.VERSION}`, JSON.stringify(storage))
 
     return next(action)
